requirements.txt
# 合同修订智能体依赖包

# 文档处理
python-docx>=0.8.11
PyPDF2>=3.0.0
pymupdf>=1.23.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# 向量嵌入和检索
sentence-transformers>=2.2.0
pinecone-client>=3.0.0

# 大模型API
openai>=1.0.0

# 图像处理
Pillow>=10.0.0
numpy>=1.24.0

# 数据可视化
matplotlib>=3.7.0

# 其他依赖
requests>=2.31.0
tqdm>=4.66.0